requirements.txt
Flask==3.0.0
markdown
bleach
markupsafe
python-dotenv
requests
Pillow
PyMySQL